requirements.txt
requests>=2.32.4
openai
psycopg2-binary
fastapi
uvicorn
sqlalchemy
jinja2>=3.1.6
python-multipart>=0.0.22
httpx
bcrypt
PyJWT
Pillow>=10.3.0
slowapi
pytest
pytest-asyncio
pytest-cov
python-dotenv
supabase
pytz
mercadopago>=2.2.0
mercadopago==2.3.0
aiosmtplib
